feat: Add support for fetching deployed models in Azure
This commit adds the functionality to fetch deployed models in Azure. It includes the implementation of the `fetchDeployedModels` function, which makes a GET request to the Azure OpenAI API to retrieve the list of deployed models. The fetched models are then returned as a JSON response.

The `handleGetModels` function has been updated to use the `fetchDeployedModels` function and return the fetched models as a response.

This feature allows users to retrieve information about the deployed models in Azure, enabling a more consistent experience to the openai api in open web-ui.